added support for generating an pdf invoice from the order page, and then delete old product images when they are edited or delete

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
   let message = req.flash('error');
@@ -159,6 +160,7 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path
       }
       product.description = updatedDescription;
@@ -177,6 +179,13 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = async (req, res, next) => {
   const prodId = req.body.productId;
   try {
+    
+    const product = await Product.findById(prodId)
+    if (!product) {
+      return next(new Error('Product not found'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+
     await Product.deleteOne({ _id: prodId, userId: req.user._id });
     console.log('Successfully deleted product');
     res.redirect('/admin/products');
